Make catalog handle local fs
