+++ tests/unit/test_dataset_metadata.py
@@ -1,30 +1,8 @@
-import requests
 import pytest
 
-from pinecone_datasets.catalog import Catalog
 from pinecone_datasets.dataset_metadata import DatasetMetadata, DenseModelMetadata
 
 from pydantic import ValidationError
-
-
-def download_json_from_https(url):
-    return requests.get(url).json()
-
-
-def test_catalog():
-    catalog = Catalog()
-    catalog.load()
-
-    catalog_as_dict = download_json_from_https(
-        "https://storage.googleapis.com/pinecone-datasets-dev/quora_all-MiniLM-L6-bm25/metadata.json"
-    )
-
-    found = False
-    for dataset in catalog.list_datasets(as_df=False):
-        if catalog_as_dict["name"] == dataset:
-            found = True
-            break
-    assert found
 
 
 def test_metadta_fields_minimal():
